+++ smartimmo/src/main/java/com/gangdestrois/smartimmo/infrastructure/rest/dto/ProspectResponse.java
@@ -2,7 +2,7 @@ package com.gangdestrois.smartimmo.infrastructure.rest.dto;
 
 import com.gangdestrois.smartimmo.domain.prospect.model.Prospect;
 
-public record ProspectResponse(Integer id) {
+public record ProspectResponse(Long id) {
     public static ProspectResponse fromModel(Prospect prospect) {
         return new ProspectResponse(prospect.getId());
     }
